+++ asdf/tests/test_schema.py
@@ -142,7 +142,7 @@ def test_schema_example(filename, example):
 
     # Add some dummy blocks so that the ndarray examples work
     for i in range(3):
-        b = block.Block(np.empty((1024*1024*8), dtype=np.uint8))
+        b = block.Block(np.zeros((1024*1024*8), dtype=np.uint8))
         b._used = True
         ff.blocks.add(b)
     b._array_storage = "streamed"
@@ -150,10 +150,7 @@ def test_schema_example(filename, example):
     try:
         # Ignore warnings that result from examples from schemas that have
         # versions higher than the current standard version.
-        # TODO: this should no longer be necessary once the library can
-        # actually account for higher versions
-        warnings.simplefilter('ignore', UserWarning)
-        ff._open_impl(ff, buff)
+        ff._open_impl(ff, buff, ignore_version_mismatch=True)
     except:
         print("From file:", filename)
         raise